requirements.txt
django==3.2.9
django-choices==1.7.1
djangorestframework==3.11.0